% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Returns the price of x USD in EUR}
\usage{
convert_usd_to_eur(x, decimals = 3)
}
\arguments{
\item{x}{the usd amount to convert to EUR}

\item{decimals}{precision of rounding in decimals}
}
\value{
integer
}
\description{
Returns the price of x USD in EUR
}
\examples{
convert_usd_to_eur(100)
convert_usd_to_eur(20, 5)
}
