% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\name{eur}
\alias{eur}
\title{Returns value in formatted with euro symbol}
\usage{
eur(x)
}
\arguments{
\item{x}{the amount to transform with eur symbol}
}
\value{
string
}
\description{
Returns value in formatted with euro symbol
}
\examples{
eur(1000)
eur(4242.4242)
}
