% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{Returns the price of 1 USD in EUR}
\usage{
eurusd(decimals = 3)
}
\arguments{
\item{decimals}{precision of rounding in decimals}
}
\value{
integer
}
\description{
Returns the price of 1 USD in EUR
}
\examples{
eurusd()
eurusd(1)
}
