% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{Returns the historical price of 1 USD in EUR}
\usage{
eurusds(date_from, date_to, decimals = 3)
}
\arguments{
\item{date_from}{first date of the examined period}

\item{date_to}{last date of the examined period}

\item{decimals}{precision of rounding in decimals}
}
\value{
integer
}
\description{
Returns the historical price of 1 USD in EUR
}
\examples{
eurusds('2019-01-01', '2019-02-01')
eurusds('2019-01-01', '2019-02-01', 5)
}
